move pydantic compat layer
[INTERNAL_BRANCH=mv-pydantic-compat-layer]
